[40] Add test for Gaurdian
