+++ lib/kwtool/account/error_handler.ex
@@ -1,16 +1,12 @@
-defmodule Kwtool.Account.Guardian do
-  use Guardian, otp_app: :kwtool
+defmodule Kwtool.Account.ErrorHandler do
+  @behaviour Guardian.Plug.ErrorHandler
 
-  alias Kwtool.Account.Users
+  use KwtoolWeb, :controller
 
-  def subject_for_token(user, _claims) do
-    {:ok, to_string(user.id)}
-  end
-
-  def resource_from_claims(%{"sub" => id}) do
-    user = Users.get_user!(id)
-    {:ok, user}
-  rescue
-    Ecto.NoResultsError -> {:error, :resource_not_found}
+  @impl Guardian.Plug.ErrorHandler
+  def auth_error(conn, {type, _reason}, _opts) do
+    conn
+    |> put_resp_content_type("text/plain")
+    |> send_resp(401, to_string(type))
   end
 end
