+++ lib/kwtool/account/guardian.ex
@@ -1,12 +1,16 @@
-defmodule Kwtool.Account.ErrorHandler do
-  @behaviour Guardian.Plug.ErrorHandler
+defmodule Kwtool.Account.Guardian do
+  use Guardian, otp_app: :kwtool
 
-  use KwtoolWeb, :controller
+  alias Kwtool.Account.Users
 
-  @impl Guardian.Plug.ErrorHandler
-  def auth_error(conn, {type, _reason}, _opts) do
-    conn
-    |> put_resp_content_type("text/plain")
-    |> send_resp(401, to_string(type))
+  def subject_for_token(user, _claims) do
+    {:ok, to_string(user.id)}
+  end
+
+  def resource_from_claims(%{"sub" => id}) do
+    user = Users.get_user!(id)
+    {:ok, user}
+  rescue
+    Ecto.NoResultsError -> {:error, :resource_not_found}
   end
 end
