+++ test/kwtool/account/error_handler_test.exs
@@ -1,0 +1,5 @@
+defmodule Kwtool.Account.ErrorHandlerTest do
+  use Kwtool.DataCase, async: true
+
+  alias Kwtool.Account.ErrorHandler
+end
